:sparkles: Feat: SpringSecurityConfig 수정, DTO 디렉토리 추가
