+++ src/main/java/roomit/web1_2_bumblebee_be/domain/member/controller/MemberController.java
@@ -2,9 +2,9 @@ package roomit.web1_2_bumblebee_be.domain.member.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-import roomit.web1_2_bumblebee_be.domain.member.request.MemberRegisterRequest;
-import roomit.web1_2_bumblebee_be.domain.member.request.MemberUpdateRequest;
-import roomit.web1_2_bumblebee_be.domain.member.response.MemberResponse;
+import roomit.web1_2_bumblebee_be.domain.member.dto.request.MemberRegisterRequest;
+import roomit.web1_2_bumblebee_be.domain.member.dto.request.MemberUpdateRequest;
+import roomit.web1_2_bumblebee_be.domain.member.dto.response.MemberResponse;
 import roomit.web1_2_bumblebee_be.domain.member.service.MemberService;
 
 @RestController
