+++ src/main/java/roomit/web1_2_bumblebee_be/domain/member/dto/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package roomit.web1_2_bumblebee_be.domain.member.response;
+package roomit.web1_2_bumblebee_be.domain.member.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
