+++ src/main/java/roomit/web1_2_bumblebee_be/domain/member/entity/Role.java
@@ -1,6 +1,6 @@
 package roomit.web1_2_bumblebee_be.domain.member.entity;
 
 public enum Role {
-    Admin,
-    User
+    ROLE_ADMIN,
+    ROLE_USER
 }
