+++ src/main/java/roomit/web1_2_bumblebee_be/global/config/SpringSecurityConfig.java
@@ -1,0 +1,65 @@
+package roomit.web1_2_bumblebee_be.global.config;
+
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import roomit.web1_2_bumblebee_be.global.config.jwt.JWTFilter;
+import roomit.web1_2_bumblebee_be.global.config.jwt.JWTUtil;
+import roomit.web1_2_bumblebee_be.global.config.jwt.LoginFilter;
+
+@Configuration
+@EnableWebSecurity
+@EnableMethodSecurity
+@RequiredArgsConstructor
+public class SpringSecurityConfig {
+
+    private final AuthenticationConfiguration authenticationConfiguration;
+    private final JWTUtil jwtUtil;
+
+    @Bean
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration configuration) throws Exception {
+
+        return configuration.getAuthenticationManager();
+    }
+
+    @Bean
+    public BCryptPasswordEncoder bCryptPasswordEncoder(){
+
+        return new BCryptPasswordEncoder();
+    }
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception{
+        http
+                .csrf(AbstractHttpConfigurer::disable) //csrf 비활성화
+                .cors(AbstractHttpConfigurer::disable) //cors 비활성화
+                .formLogin(AbstractHttpConfigurer::disable) //form 로그인 비활성화
+                // 경로별 인가 작업
+                .authorizeHttpRequests(request -> request
+                        .requestMatchers("/login", "/", "/api/v1/member/signup").permitAll()
+                        .requestMatchers("/admin").hasRole("ADMIN")
+                        .anyRequest().authenticated())
+                // 세션 설정
+                .sessionManagement((session) -> session
+                        .sessionCreationPolicy(SessionCreationPolicy.STATELESS))//무상태
+                //커스텀 로그인 필터 추가
+                .addFilterAt(new LoginFilter(authenticationManager(authenticationConfiguration),jwtUtil), UsernamePasswordAuthenticationFilter.class)
+                //JWT 필터 추가
+                .addFilterBefore(new JWTFilter(jwtUtil), LoginFilter.class);
+
+
+        return http.build();
+
+    }
+}
