+++ src/test/java/roomit/web1_2_bumblebee_be/domain/member/service/MemberServiceTest.java
@@ -5,15 +5,14 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Bean;
 import roomit.web1_2_bumblebee_be.domain.member.entity.Member;
 import roomit.web1_2_bumblebee_be.domain.member.entity.Role;
 import roomit.web1_2_bumblebee_be.domain.member.entity.Sex;
 import roomit.web1_2_bumblebee_be.domain.member.exception.MemberNotFound;
 import roomit.web1_2_bumblebee_be.domain.member.repository.MemberRepository;
-import roomit.web1_2_bumblebee_be.domain.member.request.MemberRegisterRequest;
-import roomit.web1_2_bumblebee_be.domain.member.request.MemberUpdateRequest;
-import roomit.web1_2_bumblebee_be.domain.member.response.MemberResponse;
+import roomit.web1_2_bumblebee_be.domain.member.dto.request.MemberRegisterRequest;
+import roomit.web1_2_bumblebee_be.domain.member.dto.request.MemberUpdateRequest;
+import roomit.web1_2_bumblebee_be.domain.member.dto.response.MemberResponse;
 
 import java.util.NoSuchElementException;
 
@@ -40,7 +39,7 @@ class MemberServiceTest {
                 .sex(Sex.FEMALE)
                 .pwd("1111")
                 .email("이시현@Naver.com")
-                .role(Role.Admin)
+                .role(Role.ROLE_ADMIN)
                 .phoneNumber("010-33230-23")
                 .nickName("치킨유저")
                 .build();
@@ -65,7 +64,7 @@ class MemberServiceTest {
                 .sex(Sex.FEMALE)
                 .pwd("1111")
                 .email("이시현@Naver.com")
-                .role(Role.Admin)
+                .role(Role.ROLE_ADMIN)
                 .phoneNumber("010-33230-23")
                 .nickName("치킨유저")
                 .build();
@@ -87,7 +86,7 @@ class MemberServiceTest {
                 .sex(Sex.FEMALE)
                 .pwd("1111")
                 .email("이시현@Naver.com")
-                .role(Role.Admin)
+                .role(Role.ROLE_ADMIN)
                 .phoneNumber("010-33230-23")
                 .nickName("치킨유저")
                 .build();
@@ -107,7 +106,7 @@ class MemberServiceTest {
                 .sex(Sex.FEMALE)
                 .pwd("1111")
                 .email("이시현@Naver.com")
-                .role(Role.Admin)
+                .role(Role.ROLE_ADMIN)
                 .phoneNumber("010-33230-23")
                 .nickName("치킨유저")
                 .build();
@@ -118,7 +117,7 @@ class MemberServiceTest {
                 .age(20)
                 .pwd("1111")
                 .email("이시현@Naver.com")
-                .role(Role.Admin)
+                .role(Role.ROLE_ADMIN)
                 .phoneNumber("010-33230-23")
                 .build();
 
@@ -139,7 +138,7 @@ class MemberServiceTest {
                 .sex(Sex.FEMALE)
                 .pwd("1111")
                 .email("이시현@Naver.com")
-                .role(Role.Admin)
+                .role(Role.ROLE_ADMIN)
                 .phoneNumber("010-33230-23")
                 .nickName("치킨유저")
                 .build();
@@ -159,7 +158,7 @@ class MemberServiceTest {
                 .sex(Sex.FEMALE)
                 .pwd("1111")
                 .email("이시현@Naver.com")
-                .role(Role.Admin)
+                .role(Role.ROLE_ADMIN)
                 .phoneNumber("010-33230-23")
                 .nickName("치킨유저")
                 .build();
